bugfix: user requesting residue that doesn't exist caused seggie
As carb was already part of glycoprotein it would get to end of protein and start casting the ROH as a pdbResidue. Fixes are to not add the carb to after this step, but that's not good if the bad residue request comes later in the file. Proper fix is to check if the potential glycosite residue is a protein and skipping if not.